massive refactor and clean up
