+++ backend/server/config/createToken.js
@@ -1,7 +1,6 @@
 import jwt from 'jsonwebtoken';
-
 import { devConfig } from './devConfig';
 
-export const createToken = args => (
-  jwt.sign({ id: args._id }, devConfig.JWT_SECRET)
-);
+export const createToken = args => {
+  jwt.sign({ id: args._id }, devConfig.JWT_SECRET);
+};
