+++ backend/server/config/devConfig.js
@@ -1,4 +1,4 @@
 export const devConfig = {
   DB_URL: 'mongodb://localhost/CityImpactDB',
-  JWT_SECRET: 'iuwbgu9bg98whwiv9wg98wghne908hgoin',
+  JWT_SECRET: 'asubdyficausdfnaukdfinacusfausdn'
 };
