--- backend/server/config/host_url.js
@@ -1,1 +1,0 @@
-export const HOST_URL = 'https://cityimpact.herokuapp.com';
