+++ backend/server/index.js
@@ -1,9 +1,8 @@
 import express from 'express';
 import dbConfig from './config/db';
 import middlewareConfig from './config/middleware';
-import { EventRoutes } from './modules';
+import { EventRoutes, UserRoutes } from './modules';
 import passport from 'passport';
-import UserRoutes from './modules/users/routes';
 
 const app = express();
 
@@ -11,8 +10,8 @@ dbConfig();
 
 middlewareConfig(app);
 
-app.use('/api', [EventRoutes]);
-app.use('/api', [UserRoutes]);
+app.use('/api', EventRoutes);
+app.use('/api', UserRoutes);
 
 const PORT = process.env.PORT || 3000;
 app.listen(PORT, err => {
