--- backend/server/modules/auth/controller.js
@@ -1,1 +1,0 @@
-
