--- backend/server/modules/auth/index.js
@@ -1,5 +1,0 @@
-// import { AuthRoutes } from './routes';
-//
-// export {
-//   AuthRoutes
-// };
