+++ backend/server/modules/users/model.js
@@ -1,5 +1,4 @@
 import mongoose, { Schema } from 'mongoose';
-import bcrypt from 'bcrypt-nodejs';
 
 const UserSchema = new Schema({
     email: {
@@ -15,35 +14,19 @@ const UserSchema = new Schema({
   },
   { timestamps: true },
 );
-	// facebook: {
-	// 	id: String,
-	// 	token: String,
-	// 	email: String,
-	// 	name: String
-	// }
 
-
-// UserSchema.methods.generateHash = password => {
-// 	return bcrypt.hashSync(password, bcrypt.genSaltSync(9));
-// };
-//
-// UserSchema.methods.validPassword = password => {
-// 	return bcrypt.compareSync(password, this.user.password);
-// };
-
-UserSchema.statics.findOrCreate = async function (args) {
+UserSchema.statics.findOrCreate = async function (userInfo) {
   try {
-    const user = this.findOne({
-      email: args.email,
-      fullName: args.fullName,
+    const user = await this.findOne({
+      email: userInfo.email,
+      fullName: userInfo.fullName,
     });
-
     if (!user) {
-      return await this.create(args);
+      return await this.create(userInfo);
     }
-
     return user;
   } catch (e) {
+    console.log(`ERROR FROM USERS MODEL`);
     return e;
   }
 };
