+++ backend/server/modules/users/routes.js
@@ -1,5 +1,4 @@
 import { Router } from 'express';
-
 import * as UserController from './controller';
 
 const routes = new Router();
