--- src/config/fbConfig.js
@@ -1,3 +1,0 @@
-export default {
-  APP_ID: '324070188040369'
-};
