+++ src/util/host_util.js
@@ -1,1 +1,5 @@
-export const HOST_URL = 'http://192.168.3.206:3000';
+export const HOST_URL = 'http://10.0.0.42:3000';
+
+export const fbConfig  = {
+  APP_ID: '324070188040369'
+};
